requirements.txt
Mako==1.1.5
matplotlib==3.3.4
numpy==1.18.1
Pillow==8.0.1
PyQt5<=5.15.0
PyQt5-stubs
hg+https://hg.ximc.ru/eyepoint/epcore@t-51956#egg=epcore
hg+https://hg.ximc.ru/eyepoint/ivviewer@dev-0.1#egg=ivviewer